Optimize approx.* data types

+++ src/stats/approx.c
@@ -6,11 +6,11 @@ typedef struct
     double yhigh;
     double f1;
     double f2;
-    int kind;
-    int na_rm;
+    uint8_t kind;
+    uint8_t na_rm;
 } appr_meth;
 
-static double approx1(double v, double *x, double *y, int n,
+static double approx1(double v, double *x, double *y, uint32_t n,
                       appr_meth *Meth)
 {
     /* Approximate  y(v),  given (x,y)[i], i = 0,..,n-1 */
@@ -18,8 +18,8 @@ static double approx1(double v, double *x, double *y, int n,
     if (!n)
         return NAN;
 
-    int i = 0;
-    int j = n - 1;
+    uint32_t i = 0;
+    uint32_t j = n - 1;
     /* handle out-of-domain points */
     if (v < x[i])
         return Meth->ylow;
@@ -29,7 +29,7 @@ static double approx1(double v, double *x, double *y, int n,
     /* find the correct interval by bisection */
     while (i < j - 1)
     { /* x[i] <= v <= x[j] */
-        int ij = (i + j) / 2;
+        uint32_t ij = (i + j) / 2;
         /* i+1 <= ij <= j-1 */
         if (v < x[ij])
             j = ij;
@@ -53,9 +53,9 @@ static double approx1(double v, double *x, double *y, int n,
         return (Meth->f1 != 0.0 ? y[i] * Meth->f1 : 0.0) + (Meth->f2 != 0.0 ? y[j] * Meth->f2 : 0.0);
 } /* approx1() */
 
-static void R_approxfun(double *x, double *y, int nxy,
-                        double *xout, double *yout, int nout, int method,
-                        double yleft, double yright, double f, int na_rm)
+static void R_approxfun(double *x, double *y, uint32_t nxy,
+                        double *xout, double *yout, uint32_t nout, uint8_t method,
+                        double yleft, double yright, double f, uint8_t na_rm)
 {
     appr_meth M = {0.0, 0.0, 0.0, 0.0, 0}; /* -Wall */
 
@@ -65,13 +65,13 @@ static void R_approxfun(double *x, double *y, int nxy,
     M.ylow = yleft;
     M.yhigh = yright;
     M.na_rm = na_rm;
-    int i;
+    uint32_t i;
     for (i = 0; i < nout; i++)
         yout[i] = isnan(xout[i]) ? xout[i] : approx1(xout[i], x, y, nxy, &M);
 }
 
-static double *C_Approx(double *x, double *y, int nxy, double *xout, int nout, int method,
-                        double yleft, double yright, double f, int na_rm)
+static double *C_Approx(double *x, double *y, uint32_t nxy, double *xout, uint32_t nout, uint8_t method,
+                        double yleft, double yright, double f, uint8_t na_rm)
 {
     double *yout = calloc(nout, sizeof(double));
     R_approxfun(x, y, nxy, xout, yout, nout, method, yleft, yright, f, na_rm);
@@ -81,7 +81,7 @@ static double *C_Approx(double *x, double *y, int nxy, double *xout, int nout, i
 // approx func from
 // https://github.com/wch/r-source/blob/79298c499218846d14500255efd622b5021c10ec/src/library/stats/R/approx.R
 // "linear" method
-approx_output_t approx(int n, double *x, double *y, int nout)
+approx_output_t approx(uint32_t n, double *x, double *y, uint32_t nout)
 {
     double *xout = linspace(x[0], x[n - 1], nout);
     double *yout = C_Approx(x, y, n, xout, nout, 1, NAN, NAN, 0, 1);
